step 11, b

+++ src/main/java/com/back/domain/wiseSaying/controller/WiseSayingController.java
@@ -3,11 +3,9 @@ package com.back.domain.wiseSaying.controller;
 import com.back.AppContext;
 import com.back.domain.wiseSaying.entity.WiseSaying;
 import com.back.domain.wiseSaying.service.WiseSayingService;
+import com.back.global.rq.Rq;
 
-import java.util.Arrays;
-import java.util.Map;
 import java.util.Scanner;
-import java.util.stream.Collectors;
 
 public class WiseSayingController {
     private final WiseSayingService wiseSayingService = AppContext.wiseSayingService;
@@ -34,44 +32,36 @@ public class WiseSayingController {
         }
     }
 
-    public void actionDelete(String cmd) {
-        String[] cmdBits = cmd.split("\\?", 2);
-        String queryString = cmdBits[1];
+    public void actionDelete(Rq rq) {
+        int id = rq.getParamsInt("id", -1);
 
-        Map<String, String> params = Arrays
-                .stream(queryString.split("&"))
-                .map(e -> e.split("=", 2))
-                .filter(e -> e.length == 2 && !e[0].isBlank() && !e[1].isBlank())
-                .collect(Collectors.toMap(e -> e[0].trim(), e -> e[1].trim()));
-
-        String idStr = params.getOrDefault("id", "");
-        int id = Integer.parseInt(idStr);
+        if(id == -1){
+            System.out.println("id를 정확히 입력해주세요.");
+            return;
+        }
 
         boolean deleted = wiseSayingService.delete(id);
 
-        if (!deleted) {
-            System.out.printf("%d번 명언은 존재하지 않습니다.\n", id);
+        if(!deleted){
+            System.out.printf("%d번 명언은 존재하지 않습니다.", id);
             return;
         }
+
 
         System.out.printf("%d번 명언이 삭제되었습니다.\n", id);
     }
 
-    public void actionModify(String cmd) {
-        String[] cmdBits = cmd.split("\\?", 2);
-        String queryString = cmdBits[1];
+    public void actionModify(Rq rq) {
 
-        Map<String, String> params = Arrays
-                .stream(queryString.split("&"))
-                .map(e -> e.split("=", 2))
-                .filter(e -> e.length == 2 && !e[0].isBlank() && !e[1].isBlank())
-                .collect(Collectors.toMap(e -> e[0].trim(), e -> e[1].trim()));
 
-        String idStr = params.getOrDefault("id", "");
-        int id = Integer.parseInt(idStr);
+        int id = rq.getParamsInt("id", -1);
+
+        if(id == -1){
+            System.out.println("id를 정확히 입력해주세요.");
+            return;
+        }
 
         WiseSaying wiseSaying = wiseSayingService.findById(id);
-
         if (wiseSaying == null) {
             System.out.printf("%d번 명언은 존재하지 않습니다.\n", id);
             return;
